ürün

+++ src/Controllers/OperationTrait.php
@@ -325,7 +325,10 @@ trait OperationTrait
                     foreach ($group['datas'] as $data) {
                         $relation_models[] = new $group['relation_model']($data);
                     }
-                    $this->model->$relation()->delete();
+
+                    if ($group['is_reset']) {
+                        $this->model->$relation()->delete();
+                    }
                     $this->model->$relation()->saveMany($relation_models);
                 }
                 continue;
